reduce cost from .len() ; split benches; tag 0.0.17
