+++ Cargo.toml
@@ -1,7 +1,7 @@
 
 [package]
 name = "im_ternary_tree"
-version = "0.0.16"
+version = "0.0.17"
 edition = "2021"
 authors = ["jiyinyiyong <jiyinyiyong@gmail.com>"]
 license = "MIT"
@@ -23,7 +23,7 @@ criterion = "0.3"
 
 
 [[bench]]
-name = "creating"
+name = "accessing"
 harness = false
 
 [profile.release]
